Added loader and fixed height of edited note
I added a loader div with css animation that runs until the data from
the database is loaded into the div with id "data". I also fixed some
problems with the notes when edited - the height wasn't correct. I still
have to test in mobile with the keyboard if it will drag the save and
delete buttons on top of the note's text.