Fixed countNotes variable problem
I fixed the problem where the countNotes div (which is hidden and gets
the number of notes from the DB) in some cases didn't had the right
value when the user got to the bottom of the page. Now, when that
happens, the variable gets updated again. I also fixed the problem of
the first 10 notes displayed: if the user deleted the 10 notes the page
realoaded, it wouldn't load anymore notes and none would be shown. I
fixed by setting the delete function to load 10 notes everytime the user
deletes one note.