Deleted showMoreNotes function
There was a problem with the notes in which when the user deleted a note
that wasn't in the 10 when the page is loaded, only the 10 were
displayed after that. To do that, I added a limit variable in the
showNotes and deleted the showMoreNotes function because it was
ambiguous.