Added logout button to the side bar fixed issues
The animation in the textarea that adds new notes wasn't working, but
it's already fixed. I also increased the width of the edit_note_content
in the modal.Added logout button to the side bar fixed issues

The animation in the textarea that adds new notes wasn't working, but
it's already fixed. I also increased the width of the edit_note_content
in the modal.